It's a piece of shit, but resuming works now

+++ shared/unloader.lua
@@ -1,12 +1,14 @@
 local mover = require("mover")
-local locator = require("locator")
 
 local unloader = {}
 
-local total_unloading_chests = 0
-local unloading_pos
+unloader.get_unloading_chests = function(progress)
+    return progress.unloading_chests
+end
 
-unloader.create_unloading_area = function(start_x, start_y, start_z)
+unloader.create_initial_unloading_area = function(start_x, start_y, start_z)
+    mover.move_to(start_x, start_y, start_z)
+
     mover.turn_to_direction("east")
 
     mover.move_forward()
@@ -51,22 +53,23 @@ unloader.create_unloading_area = function(start_x, start_y, start_z)
         turtle.select(2)
         turtle.placeDown()
         turtle.select(1)
-
-        total_unloading_chests = total_unloading_chests + 1
     end
 
-    unloading_pos = { x = start_x, y = start_y, z = start_z + 1 }
+    local unloading_pos = { x = start_x, y = start_y, z = start_z + 1 }
     mover.move_to(start_x, start_y, start_z)
+
+    return 1, unloading_pos
 end
 
-unloader.unload_at_chest = function(chest_number)
-    if not unloading_pos then
-        error("No unloading chests set.")
+unloader.unload_at_chest = function(chest_number, progress)
+    local chest = progress.unloading_chests[chest_number]
+    if not chest then
+        error("Could not determine unloading chest")
     end
 
-    mover.move_to_y(unloading_pos.y)
-    mover.move_to_x(unloading_pos.x)
-    mover.move_to_z(unloading_pos.z)
+    mover.move_to_y(chest.y)
+    mover.move_to_x(chest.x)
+    mover.move_to_z(chest.z)
 
     mover.turn_to_direction("east")
 
@@ -75,10 +78,23 @@ unloader.unload_at_chest = function(chest_number)
     end
 
     for i = 3, 16 do
-        turtle.select(i)
-        turtle.drop()
+        if turtle.getItemCount(i) > 0 then
+            turtle.select(i)
+            turtle.drop()
+        end
     end
     turtle.select(1)
+end
+
+unloader.should_unload = function()
+    local amount_of_filled_slots = 0
+    for i = 3, 16 do
+        if turtle.getItemCount(i) > 0 then
+            amount_of_filled_slots = amount_of_filled_slots + 1
+        end
+    end
+
+    return amount_of_filled_slots == 14
 end
 
 return unloader
